Initial support for flashing Zephyr boards OTA
This is a component that enables required configs for flashing
ESPHome projects with a zephyr platform over the air. The way
this is handled is different enough from the ESP OTA component
that I created a new one. It may be mreged at some point in the
future.

Some refactoring of this component is still pending (merging code
from the main zephyr component into this one)