SPI component working
This is not really how I would like to make different classes of
SPI components (macro gaurds and different implementations) but
I could not really see a good way around it, and still keep a
stable api that other parts of the code might depend on.

The problem is that because all the methods are templated, there
cannot be a virtual base class to use as a bass class pointer in
places like SPIDevice. Virtual tables and templates cannot be used
together, and compiling SPIDevice required that the defintion for
all the templated functions be available.

I tried just having two classes that share a base class and not
have SPIDevice look at the baseclass, but look at either class
based on ifdefs, but that would mean other parts of the code base
that try and use a baseclass pointer would run into the same
compile time issue, and thus the ifdef gaurds would propagate
around the code base. This seems like a good (if not ideal)
alternative. As long as the two classes share the same public
interface everything will "just work"

The danger is someone adds a method to one, and not the other,
but by the structure of the code I would hope an editor would
realize that there are two alternative implementations.